requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
sqlite-utils

pdfplumber
python-docx
python-pptx
openpyxl
pytesseract
Pillow

langchain
requests
